Немного рефакторингв
