+++ app/src/main/java/com/azmetov/playlistmaker/network/PlaylistResponseDto.kt
@@ -11,4 +11,17 @@ data class PlaylistResponseDto(
     val results: List<TrackDto>
 )
 
+data class TrackDto(
 
+    @SerializedName("trackTimeMillis")
+    val trackTimeMillis: Int? = null,
+
+    @SerializedName("artworkUrl100")
+    val artworkUrl100: String? = null,
+
+    @SerializedName("artistName")
+    val artistName: String? = null,
+
+    @SerializedName("trackName")
+    val trackName: String? = null
+)
